Front-End Fixed +  Post Put
Changed the front end so now users can see their balance.. also added a link to the editpage. In the edit page you can now create a user and deactive it.  for the server stub... a post method for creating new accounts is made and deactivating an account is possible, furthermore some bug fixes... good day